refactor: remove unused epubjs dependency and update atom exports
