--- packages/jotai/src/atoms/index.ts
@@ -1,1 +1,0 @@
-export * from "./kiwi";
